+++ tests/unit/snap/test_span_targets.cpp
@@ -173,6 +173,7 @@ public:
     void recordFreeGeometry(const QString&, const QString&, const QRect&, bool) override
     {
     }
+    using PhosphorEngine::IWindowTrackingService::clearFreeGeometry;
     void clearFreeGeometry(const QString&) override
     {
     }